fix buttons to use border-image instead
I ran into the issue on safari that it wasn't working correctly, so I decided to make set width and height on the vector images, and then use border-image to make the border image. I also found out how to use filter: drop-shadow() for the shadow, instead of a separate image.